Renamed driver from PhpAmqp to Amqp
